requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.0
sse-starlette>=2.0
claude-code-sdk
PyYAML>=6.0
numpy>=1.24.0
